requirements.txt
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
jmespath==1.0.1
loguru==0.7.2
mutagen==1.47.0
pycryptodomex==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
websockets==12.0
yt-dlp==2024.7.16
